+++ tests/configs/config_ds000248_no_mri.py
@@ -17,4 +17,5 @@ ch_types = ['meg']
 use_maxwell_filter = False
 process_er = False
 
-use_template_mri = True
+use_template_mri = "fsaverage"
+adjust_coreg = True